Implemented a type for LinkConnections Refactoring
